added remove food functionality

+++ MainMenu.cpp
@@ -142,18 +142,23 @@ void MainMenu::addFood() {
 
 
 void MainMenu::removeFood() {
-    // Placeholder function for removing a food
-    std::cout << "Remove Food - To be implemented\n";
-
-    //debug line
-    
     std::string foodID;
-    std::cout << "Enter the ID of the food that you want to remove: ";
+    std::cout << "Enter the food ID of the food to remove from the menu: ";
     std::cin >> foodID;
+    std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n');  // Clear the input buffer
 
-    foodList -> deleteByID(foodID);
+    Node* itemNode = foodList->searchByID(foodID);  // Utilizing searchByID to find the node
+    if (itemNode != nullptr) {
+        // Access the FoodItem data from the Node
+        FoodItem* item = &itemNode->data;
+        std::cout << "\"" << item->id << " - " << item->name << " - " << item->description
+                  << "\" is being removed from the system." << std::endl;
 
-
+        foodList->deleteByID(foodID);  // Utilizing deleteByID to remove the item
+        // No need to delete the FoodItem* as it is not dynamically allocated separately
+    } else {
+        std::cout << "No item found with ID " << foodID << std::endl;
+    }
 }
 
 void MainMenu::displayBalance() {
